+++ codalab/apps/web/templates/web/worksheets/detail.html
@@ -139,7 +139,7 @@
     <script src="{{ STATIC_URL }}bower_components/jsonapi-datastore/dist/jsonapi-datastore.min.js"></script>
 
     {# ------- JSX FILES  ------- #}
-    <script src="{{ STATIC_URL }}dist/search_bar.js"></script>
+    <script src="{{ STATIC_URL }}dist/new_main.js"></script>
     <script src="{{ STATIC_URL }}dist/button.js"></script>
     <script src="{{ STATIC_URL }}dist/context_menu.js"></script>
     <script src="{{ STATIC_URL }}dist/worksheet_utils.js"></script>
